pyproject.toml
[project]
name = "loguetown"
version = "0.1.0"
description = "Local-first multi-agent coding assistant"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "pyyaml>=6.0",
]

[project.scripts]
lt = "core.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["core*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
